Share compiler-backed debug tests
Run shared lit fixtures with solc, Solar, or both using direct CLI
outputs instead of the Python preparation adapter. Keep common checks
and compiler-specific expectations, including intentional unknowns.

Run the shared suite in compiler CI jobs and document independent modes.
