--- test/solar/capabilities/bytes-length.test
@@ -1,47 +1,0 @@
-# REQUIRES: soldb
-# RUN: rm -rf %t
-# RUN: %prepare %S/../Dynamic.sol Dynamic --output %t --optimization %optimization --checkpoints bytes-length
-
-# Check both executions before comparing their debug information.
-# RUN: %soldb run %t/solar/Dynamic.bin "size(bytes)" "0x010203" --save-trace %t/solar.trace.json --json > %t/solar.run.json
-# RUN: %soldb run %t/solc/Dynamic.bin "size(bytes)" "0x010203" --save-trace %t/solc.trace.json --json > %t/solc.run.json
-# RUN: jq -c '[.success, .output]' %t/solar.trace.json %t/solc.trace.json | FileCheck %s --check-prefix=EXEC
-# EXEC: [true,"0x0000000000000000000000000000000000000000000000000000000000000003"]
-# EXEC-NEXT: [true,"0x0000000000000000000000000000000000000000000000000000000000000003"]
-
-# NOTE: Gas-mode tail merging shares this return with sum's implicit return,
-# and its incoming jump becomes fallthrough. Neither format has an unambiguous
-# source checkpoint. Keep codegen unchanged and require an inconclusive report,
-# not a guessed location or vacuous equivalence. Other modes retain the checkpoint.
-# RUN: %if optimization-gas %{ not %} %soldb debug-diff --reference-trace-file %t/solar.trace.json --candidate-trace-file %t/solar.trace.json --reference-ethdebug-dir 0x5fbdb2315678afecb367f032d93f642f64180aa3:Dynamic:%t/solar --candidate-ethdebug-dir 0x5fbdb2315678afecb367f032d93f642f64180aa3:Dynamic:%t/ethdebug --mode coverage --json > %t/formats.diff.json
-# RUN: jq -c '[.executionEquivalent, .comparable, .equivalent]' %t/formats.diff.json | FileCheck %s %if optimization-gas %{ --check-prefix=UNKNOWN %} %else %{ --check-prefix=DIFF %}
-# RUN: %if optimization-gas %{ jq -c '[.reference.mappedInstructions, .candidate.mappedInstructions, .reference.sourceSteps, .candidate.sourceSteps, .differenceCount, .diagnostics]' %t/formats.diff.json | FileCheck %s --check-prefix=UNKNOWN-FORMATS %}
-# UNKNOWN-FORMATS: [0,0,0,0,0,["reference trace has no source steps","candidate trace has no source steps"]]
-
-# The solc checkpoint remains required; gas mode diagnoses only the missing candidate stop.
-# RUN: %if optimization-gas %{ not %} %soldb debug-diff --reference-trace-file %t/solc.trace.json --candidate-trace-file %t/solar.trace.json --reference-ethdebug-dir 0x5fbdb2315678afecb367f032d93f642f64180aa3:Dynamic:%t/solc --candidate-ethdebug-dir 0x5fbdb2315678afecb367f032d93f642f64180aa3:Dynamic:%t/ethdebug --mode coverage --checkpoints-file %t/checkpoints.json --json > %t/solc.diff.json
-# RUN: jq -c '[.executionEquivalent, .comparable, .equivalent]' %t/solc.diff.json | FileCheck %s %if optimization-gas %{ --check-prefix=UNKNOWN %} %else %{ --check-prefix=DIFF %}
-# RUN: %if optimization-gas %{ jq -c '[(.reference.sourceSteps > 0), .candidate.mappedInstructions, .candidate.sourceSteps, .differenceCount, .diagnostics]' %t/solc.diff.json | FileCheck %s --check-prefix=UNKNOWN-SOLC %}
-# UNKNOWN-SOLC: [true,0,0,1,["candidate: checkpoint Dynamic.sol:6 was not reached","candidate trace has no source steps"]]
-# UNKNOWN: [true,false,false]
-# DIFF: [true,true,true]
-
-# Both formats account for all program gas and still produce a flamegraph.
-# In gas mode, legacy attribution stays unknown. ETHDebug may retain alternatives
-# that the profiler resolves with compiler-authored function identity; any source
-# attribution must belong to the executed checkpoint, never sum's shared return.
-# RUN: %soldb profile --trace-file %t/solar.trace.json --ethdebug-dir 0x5fbdb2315678afecb367f032d93f642f64180aa3:Dynamic:%t/ethdebug --flamegraph %t/ethdebug.svg --json > %t/ethdebug.profile.json
-# RUN: %soldb profile --trace-file %t/solar.trace.json --ethdebug-dir 0x5fbdb2315678afecb367f032d93f642f64180aa3:Dynamic:%t/solar --flamegraph %t/source-maps.svg --json > %t/source-maps.profile.json
-# RUN: %if !optimization-gas %{ jq -c '[.totals.unmappedGas, (.totals.sourceGas > 0)]' %t/ethdebug.profile.json %t/source-maps.profile.json | FileCheck %s --check-prefix=PROFILE %}
-# RUN: %if optimization-gas %{ jq -c '[.totals.unmappedGas, (.totals.programGas > 0), (.totals.sourceGas + .totals.sourcelessGas == .totals.programGas), (.totals.programGas == .totals.stepGas)]' %t/ethdebug.profile.json %t/source-maps.profile.json | FileCheck %s --check-prefix=ACCOUNTING %}
-# RUN: %if optimization-gas %{ jq -c '[.totals.sourceGas, (.totals.sourcelessGas == .totals.programGas), (.sourceLines | length)]' %t/source-maps.profile.json | FileCheck %s --check-prefix=SOURCELESS %}
-# RUN: %if optimization-gas %{ jq -c --slurpfile checkpoints %t/checkpoints.json '[all(.sourceLines[]; .source == $checkpoints[0][0].source and .line == $checkpoints[0][0].line), (([.sourceLines[].gas] | add // 0) == .totals.sourceGas)]' %t/ethdebug.profile.json | FileCheck %s --check-prefix=ETHDEBUG-SOURCE %}
-# ACCOUNTING: [0,true,true,true]
-# ACCOUNTING-NEXT: [0,true,true,true]
-# SOURCELESS: [0,true,0]
-# ETHDEBUG-SOURCE: [true,true]
-# PROFILE: [0,true]
-# PROFILE-NEXT: [0,true]
-# RUN: FileCheck %s --check-prefix=SVG < %t/ethdebug.svg
-# RUN: FileCheck %s --check-prefix=SVG < %t/source-maps.svg
-# SVG: <svg
