--- test/solar/capabilities/helper.test
@@ -1,29 +1,0 @@
-# REQUIRES: soldb
-# RUN: rm -rf %t
-# RUN: %prepare %S/../ControlFlow.sol ControlFlow --output %t --optimization %optimization --checkpoints helper
-
-# Check both executions before comparing their debug information.
-# RUN: %soldb run %t/solar/ControlFlow.bin "internalCall(uint256)" "3" --save-trace %t/solar.trace.json --json > %t/solar.run.json
-# RUN: %soldb run %t/solc/ControlFlow.bin "internalCall(uint256)" "3" --save-trace %t/solc.trace.json --json > %t/solc.run.json
-# RUN: jq -c '[.success, .output]' %t/solar.trace.json %t/solc.trace.json | FileCheck %s --check-prefix=EXEC
-# EXEC: [true,"0x0000000000000000000000000000000000000000000000000000000000000007"]
-# EXEC-NEXT: [true,"0x0000000000000000000000000000000000000000000000000000000000000007"]
-
-# Legacy maps and ETHDebug must agree on the same machine trace.
-# RUN: %soldb debug-diff --reference-trace-file %t/solar.trace.json --candidate-trace-file %t/solar.trace.json --reference-ethdebug-dir 0x5fbdb2315678afecb367f032d93f642f64180aa3:ControlFlow:%t/solar --candidate-ethdebug-dir 0x5fbdb2315678afecb367f032d93f642f64180aa3:ControlFlow:%t/ethdebug --mode coverage --json > %t/formats.diff.json
-# RUN: jq -c '[.comparable, .equivalent]' %t/formats.diff.json | FileCheck %s --check-prefix=DIFF
-
-# Explicit checkpoints must survive in both compilers, even if both formats lost them.
-# RUN: %soldb debug-diff --reference-trace-file %t/solc.trace.json --candidate-trace-file %t/solar.trace.json --reference-ethdebug-dir 0x5fbdb2315678afecb367f032d93f642f64180aa3:ControlFlow:%t/solc --candidate-ethdebug-dir 0x5fbdb2315678afecb367f032d93f642f64180aa3:ControlFlow:%t/ethdebug --mode coverage --checkpoints-file %t/checkpoints.json --json > %t/solc.diff.json
-# RUN: jq -c '[.comparable, .equivalent]' %t/solc.diff.json | FileCheck %s --check-prefix=DIFF
-# DIFF: [true,true]
-
-# Both formats must attribute gas and produce a flamegraph.
-# RUN: %soldb profile --trace-file %t/solar.trace.json --ethdebug-dir 0x5fbdb2315678afecb367f032d93f642f64180aa3:ControlFlow:%t/ethdebug --flamegraph %t/ethdebug.svg --json > %t/ethdebug.profile.json
-# RUN: %soldb profile --trace-file %t/solar.trace.json --ethdebug-dir 0x5fbdb2315678afecb367f032d93f642f64180aa3:ControlFlow:%t/solar --flamegraph %t/source-maps.svg --json > %t/source-maps.profile.json
-# RUN: jq -c '[.totals.unmappedGas, (.totals.sourceGas > 0)]' %t/ethdebug.profile.json %t/source-maps.profile.json | FileCheck %s --check-prefix=PROFILE
-# PROFILE: [0,true]
-# PROFILE-NEXT: [0,true]
-# RUN: FileCheck %s --check-prefix=SVG < %t/ethdebug.svg
-# RUN: FileCheck %s --check-prefix=SVG < %t/source-maps.svg
-# SVG: <svg
